+++ agent/custom/action/Common/click.py
@@ -1,23 +1,10 @@
-import json
-
-from .utils import click_rect
+from .utils import click_rect, load_params
 
 from maa.agent.agent_server import AgentServer
 from maa.custom_action import CustomAction
 from maa.context import Context
+from utils.logger import logger
 
-
-def load_params(custom_action_param) -> dict:
-    """兼容 None / dict / JSON 字符串（含 "null"、"{}"）的 custom_action_param 解析。"""
-    if not custom_action_param:
-        return {}
-    if isinstance(custom_action_param, dict):
-        return custom_action_param
-    try:
-        params = json.loads(custom_action_param)
-    except Exception:
-        return {}
-    return params if isinstance(params, dict) else {}
 
 @AgentServer.custom_action("click_override")
 class ClickOverride(CustomAction):
@@ -27,7 +14,7 @@ class ClickOverride(CustomAction):
         params = load_params(argv.custom_action_param)
         target = params.get("target")
 
-        if target and len(target) == 4:
+        if isinstance(target, (list, tuple)) and len(target) == 4:
             click_rect(controller, target, 0.005)
             return CustomAction.RunResult(success=True)
 
@@ -35,4 +22,10 @@ class ClickOverride(CustomAction):
             click_rect(controller, argv.box, 0.005)
             return CustomAction.RunResult(success=True)
 
+        logger.warning(
+            "ClickOverride: no valid click target. "
+            "custom_action_param=%r, target=%r",
+            argv.custom_action_param,
+            target,
+        )
         return CustomAction.RunResult(success=False)
